Rename MainArguments
